backend/kafis/requirements.txt
amqp==2.5.2
astroid==2.3.2
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
Django==2.2.7
django-cors-headers==3.1.1
django-extensions==2.2.5
djangorestframework==3.10.3
flower==0.9.3
importlib-metadata==0.23
isort==4.3.21
kombu==4.6.6
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
Pillow==6.2.1
psycopg2==2.8.4
pydotplus==2.0.2
pylint==2.4.3
pyparsing==2.4.2
pytz==2019.3
redis==3.3.11
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
typed-ast==1.4.0
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
scipy
pandas
scikit-learn
nose
nolearn
scikit-image
opencv-python
dlib
Keras
tensorflow